Implement a wrapper in make so common.sh is only calculated once
Wrapper is a hack that sources common.sh before running `make all` to
ensure that derived variables are calculated and exported once. This is
especially useful for variables calculated by hitting an external API.
Instead of hitting the API in each script as part of the install, it'll
only happen here once.

  - `make` runs wrapper.sh
  - wrapper.sh sources logging.sh to ensure we log the next step
  - wrapper.sh sources common.sh
  - wrapper.sh restores stdout/stderr to the tty, so each additional
  script is logged independently
  - wrapper.sh calls `make all`, which runs each of the numbered
  installation scripts